+++ pulsar/apps/data/redis/client.py
@@ -1,5 +1,4 @@
 from itertools import chain
-from hashlib import sha1
 import datetime
 
 import pulsar