+++ pulsar/apps/ds/__init__.py
@@ -3,3 +3,10 @@ from .client import COMMANDS_INFO, redis_to_py_pattern
 from .parser import (PyRedisParser, RedisParser, redis_parser,
                      RedisError, ResponseError,
                      InvalidResponse, NoScriptError, CommandError)
+
+
+__all__ = ['PulsarDS', 'DEFAULT_PULSAR_STORE_ADDRESS', 'pulsards_url',
+           'COMMANDS_INFO', 'redis_to_py_pattern',
+           'PyRedisParser', 'RedisParser', 'redis_parser',
+           'RedisError', 'ResponseError',
+           'InvalidResponse', 'NoScriptError', 'CommandError']